Added in an option to force the upload of the Firmata sketch.
This version of PyMata checks to see if it can connect to Firmata, and uploads
the FirmataPlus sketch if it cannot connect. However, currently it's hard
if not impossible to confirm that you've got the correct Firmata sketch
loaded. You could check the Firmata protocol version, but there may be a
number of different sketches which implement the same protocol (yes, modified
versions probably should be given their own firmware version). Anyway, until
I can think of a better solution, it's now possible to pass a
'force_programming' parameter to PyMata which automatically triggers an
attempt to upload the FirmataPlus sketch.